move category widget to folder of category
