+++ lib/features/home/presentation/views/home_view.dart
@@ -7,7 +7,7 @@ import 'package:shop_app/features/home/presentation/widgets/banners/banners_list
 import '../widgets/cart_head_icon.dart';
 import '../widgets/custom_search_bar.dart';
 import '../widgets/custom_staggered_gride_view.dart';
-import '../widgets/category_text.dart';
+import '../widgets/category/category_text.dart';
 import '../widgets/category/list_of_category.dart';
 import '../widgets/see_all_text.dart';
